Add files via upload
EDA plays a crucial role in the data analysis process by providing a comprehensive understanding of the data and guiding further analysis. It helps to uncover patterns, relationships, and outliers that may not be immediately apparent, and provides a solid foundation for making informed decisions and drawing meaningful conclusions from the data.

Signed-off-by: YekinData-Analysis <144348733+YekinData-Analysis@users.noreply.github.com>